Split wagtaildocs tests into multiple files
